Normalize parameters to AdbClientSocket::CommandCallback parameters
The 1st parameter (result) currently has unclear meaning. Sometimes it is used 
as a network result (such as net::OK) and sometimes as the offset of the content
in the 2nd parameter (response).

This patch removes the confusion. result is always network error code and the response
is always the response content (empty if result is an error)

BUG=None

Review URL: https://chromiumcodereview.appspot.com/22429005