Make sure that 30% of restored window is always visible.
WorkspaceLayoutManager::UpdateBoundsFromShowState change

 * Make sure the restore bounds fits to the workspace.
 * Minimized window doesn't have restore bounds, so
   use the window bounds to adjust the bounds.

AdjustBoundsToEnsureWindowVisibility
 * Do not check Intersects because we always should make
   min width/height visible.

define separate AdjustWindowsBoundsWhenAdded

I also make the minimum percentage to 30% of width/height
because 66% seems to be too much. This change is short term.
This will be changed so that we keep relative position to
the old display in new display.

BUG=249043,264044,248961,263862
TEST=covered by test
R=jamescook@chromium.org, skuhne@chromium.org

Review URL: https://codereview.chromium.org/21979005