Ensure cleanup after Android systrace
On end of Android tracing, EnableIncludedCategoryGroups() is called
to cleanup the enabled categories. However, it doesn't do what is
expected when enable_count_==0, causing the category groups are left
enabled.

Now let the function consider enabled_count_, and renamed it to
UpdateCategoryGroupEnabledFlags(), so that it can also be used to
cleanup the enabled flags.

BUG=267580

Review URL: https://chromiumcodereview.appspot.com/22496002