Remove a 'content-updated' event in DirectoryTree on Files.app
A 'content-updated' event was used to notify a change of the directory tree content, in order to hide the tree if the current volume doesn't have any folder. But now, the tree is only for Google Drive volume, which always has a folder (My Drive, Offline, Recent, ...), so that it's never hidden by 'content-update' event. It's unnecessary now, and we can remove it.

BUG=269903
TEST=manual
R=mtomasz@chromium.org

Review URL: https://codereview.chromium.org/22663002