Add drive::FileSystem::GetFileByPathForSaving().
This is a part of the effort to enable Drive in the Save-As dialog of
Chrome OS for every call site. For that, we need to care about existing
callers that just simply write to local file paths returned by the dialog.

By calling the newly added method, Drive file system will return a local
cache whose modification is monitored for syncing.

BUG=140425

Review URL: https://chromiumcodereview.appspot.com/22335004