Correct the flag: is_extension_upgrade, which denote whether the extension is being upgrading.
In the function BrowserActionsContainer::BrowserActionAdded of chrome/browser/ui/views/browser_actions_container.cc, if the extension is marked as upgrading, it will stop enlarge the container if it was already at maximum size. The upgrading extension didn't remove the browser action, so it works correctly. However, when a extension is crashed (The browser action will be removed if it has) and reloaded (will  add the browser action icon back if it has), it shouldn't be marked as upgrading. The bug is introduced by http://src.chromium.org/viewvc/chrome?revision=196634&view=revision. 

BUG=261996
TEST=On Windows, reload a crashed extension shouldn't resize the browser action container of browser. For more detail, please refer to the bug.

Review URL: https://chromiumcodereview.appspot.com/20909002