Remove high-frequency trace events.
PixelBufferRasterWorkerPool::OnRasterTaskCompleted() has interesting
information traced but seriously distorts the overall timing of
the trace on Android devices. These traces should only be enabled
when the debugging information is necessary.

This patch also adds the count of tasks to
WorkerPool::ProcessCompletedTasks() as a partial replacement, but
that can probably be skipped.

R=reveman@chromium.org
BUG=268998

Review URL: https://chromiumcodereview.appspot.com/22439002