NSS: move the first protocol to the end of the ALPN extension.
NPN defines the first protocol in the client's list to be the protocol that the
client will select when there's no overlap - i.e. it's the fallback protocol.
However, then the NPN data is used for ALPN, it was becoming the most
preferable protocol because ALPN says that they are in preference order.

This change causes NSS to be aware that the data that is set by
SSL_SetNextProtoNego is in NPN's order and thus the first element should be
moved to the end when sending an ALPN extension.

BUG=267858

Review URL: https://chromiumcodereview.appspot.com/22237002