Land Recent QUIC changes.
Minor logging improvement to FEC revived packets.

Merge internal change: 50495854

Fixed existent typo: existant becomes existent.

Merge internal change: 50439145

Fix a bug in quic_connection where an Ack could be sent in response to
an ack before the packet was registered as having been received in
OnPacketComplete(), causing an entropy to be generated which was
incorrect.

Merge internal change: 50396820

Added a TODO to support "Rate limiting a warning log."

Merge internal change: 50395138

Fixing two bug where we could kill off a session twice, one of which
could "easily" happen, and one of which should only ever happen with
horrible configuration.

The first, is processing packets via MaybeProcessUndecryptablePackets or
MaybeProcessRevivedPacket after the connection had disconnected.  The
likliest way for this to happen would be to lose the CHLO and have two
'bad' packets buffered such that we'd loop on both, both would cause a
connection close which would get passed to the session and the
dispatcher, causing unhappiness.

The second is kind of silly and caught by the DCHECK I added to detect
double-disconnect: when we register idle alarms we check for timeouts.
In the case of our timeouts test we set the timeout so small the
session is basically timed out before it's created.  In this case when
we create the session we'd close it when registering the first timeout
and close again registering the second.  This should only ever be hit
if you have a horrible config or a 10 second machine hiccup.

Merge internal change: 50393796

Actually waiting for server SHLO before upping packet loss.

Merge internal change: 50384306

Fixing a crashing bug in quic_session: if a pointer is NULL maybe we
shouldn't dereference it...

Merge internal change: 50378549

Remove mostly unused set_max_open_streams method from QuicSession.

Merge internal change: 50275870

Reducing/removing test flakeiness due to SHLO being particularly
important.

Merge internal change: 50272102

Small formatting fix.

Merge internal change: 50262875

Fixing a bug in protocol negotiation where the verion flag would always
be sent.

Basically if we immediately transition to NEGOTIATED_VERSION, we never
enter the if block in OnPacketHeader and call StopSendingVersion

Merge internal change: 50182524

Added max_tcp_congestion_window_ to TcpCubicSender for testing.


Decrease the maximum congestion window to 50 from 10000 until the RTO
logic is improved.  The intent is to handle cases when the RTT is
larger than the RTO specially and not retransmit agressively then.

Merge internal change: 50080104

Change kMaxRetransmissionTimeMs to 60 seconds (from 10 seconds) to
match TCP.

Merge internal change: 50073957

Add comments to point out it is OK to specify both kX509 and kX59R, and
to document the new ecdsa_ok parameter of ProofSource::GetProof.

Merge internal change: 50071950

Do not send acks in response to acks.

Merge internal change: 50064954

R=rch@chromium.org

Review URL: https://chromiumcodereview.appspot.com/22661002