when dragging from textfield, draw text without subpixel aa and in black
When dragging text from a views textfield on Windows, we have to make
sure not to use subpixel AA (because it'll be rendered transparently
when dragged), and we need to override the theme color to render in
black (otherwise, the text without selection background will be white).

Neither of these happen to show up on Linux because render_text_linux.cc
always has subpixel AA off, and the theme leaves the text as black
whether it's selected or not.

This only fixes the views textfield dragging mentioned in the linked
bug. Unfortunately, the omnibox dragging boldness must be caused by
something else.

R=msw@chromium.org
BUG=224425

Review URL: https://chromiumcodereview.appspot.com/22477004