Add MediaStreamMsg_StopGeneratedStream message.
Use the new message to indicate that user has requested to stop stream, and use
the MediaStreamMsg_StreamGenerationFailed to indicate that browser has failed
to generate a stream.

It is not correct to use the MediaStreamMsg_StreamGenerationFailed message to
stop a stream that is already generated, because the renderer is unable to
look up the request by |request_id| once the stream is already generated.

BUG=267899
TEST=trybot, locally verified with desktop audio mirroring

Review URL: https://chromiumcodereview.appspot.com/22719002