Add a pref for a one-time wipe of gesture prefs
Because of a bug in the gesture config UI (chrome://gesture), whenever a
user tried to reset a gesture preference to default, the preference was
set to whatever the default value of the preference was at that time,
instead of unsetting the preference. This bug is now fixed, but many
users may have unintentionally set their gesture preferences to the
default values.

A temporary preference is introduced to unset all gesture preferences
once.  This preference will be removed in a proper time in future when
we think that a good portion of users have their gesture preferences
unset.

BUG=251858

Review URL: https://chromiumcodereview.appspot.com/21983002