Fix pinch gestures with nonscrollable root layer.
If the root scroll layer is nonscrollable (particularly when fully
zoomed out on a viewport-sized page), then ScrollIgnored is returned on
ScrollBegin, but we still want to support the scrolling portion of pinch
gestures.

Fix this by forcing CurrentlyScrollingLayer to the root layer
on pinch (which is also desirable so that the scroll portion doesn't go
to sublayers anyway).  Secondly, always call ScrollEnd so that the pinch-set
CurrentlyScrollingLayer is cleared.

New test case in LayerTreeHostImplTest.PinchGesture.

NOTRY=true
BUG=269669

Review URL: https://chromiumcodereview.appspot.com/22527005