Fix slow playback of 1-channel MP3 and AAC files for WebAudio on Android
BUG=266006

If the MediaFormat indicates 1 channel for the file but the MediaCodec decoder changes the output format to 2 channels, we remove the data from the second channel before sending to the receiver.  This makes the the transmitted data consistent with the expected number of channels.

Review URL: https://chromiumcodereview.appspot.com/21618002