Request profiles info in BrowserOptions.initializePage to populate the profiles list
Profiles info is being passed to JS in |BrowserOptionsHandler::GetLocalizedValues| which gets called only once when the settings page is opened but not on page reload.
When a user edits in the profiles list (adds or deletes a profile) and then reload the settings page we end up populating the profiles list with old cached data. Accordingly,
we request the profiles info in BrowserOptions.initializePage to get a fresh list of the profiles info every time the page is loaded.

BUG=269942

Review URL: https://chromiumcodereview.appspot.com/22642011