[NaCL SDK] Ship sys/mount.h for newlib/pnacl.
The primary motivation for this change was to remove the
unnecessary include of kernel_wrap.h from nacl_io.h.

In doing this is realised that we shouldn't be declaring
the mount(2) function in nacl_io.h but instread relying
on <sys/mount.h> for this.

The doxygen comment that was previously attached to
the mount(2) declaration was merged with nacl_io_init.

R=binji@chromium.org

Review URL: https://codereview.chromium.org/22642016