Process incoming updates and deletes properly.
There are two paths into synced notifications - MergeDataAndStartSyncing, and
ProcessSyncChanges.

We had been handling updates and deletes on the MergeDataAndStartSyncing (catch
up pull path) case, but not in the ProcessSyncChanges (push path).  This change
adds the missing code for the push path and some unit tests.

We also refactor slightly the Display method, replacing calls with
"UpdateInMessageCenter", which will display or remove as needed.

BUG=270175

Review URL: https://chromiumcodereview.appspot.com/22470006