Fix scoped of closer for  ACK for editor-related changes.
Sending the ACK for for editor-related changes not originating from
browser uses a ScopedClosureRunner. However, since this was defined
within an if-block, the scope ended immediately.

This changes it to have a scope of the whole method, and instead
releasing it early if we are not supposed to send the ACK.

BUG=145521, 230848, 172845, 248544, 235704

Review URL: https://chromiumcodereview.appspot.com/22411008