[NaCl SDK] Fix PNACL for debugging.
make TOOLCHAIN=pnacl DEBUG is broken.

This CL redirects all nacl-gdb calls to the gdb in newlib.
For DEBUG builds, the EXECUTABLES=*.nexe
For RELESE builds, the EXECUTABLES=*.pexe

This creates a dependency on the *.bc -> *.nexe translation and
feeds the output the the NMF generator, making the debugging case
essitial the same as newlib.

For RELEASE configs, we build a finalized .pexe and load that.

common.js now only use application-pnacl for release versions.

BUG=270745
R=sbc@chromium.org

Review URL: https://codereview.chromium.org/22766003