Refactor the client certificate code in chromeos/network/.
This prepares for the next commit, which adds automatic resolution of client certificates which doesn't require a manual trigger by the user anymore.

The only functional is that flimflam::kOpenVPNClientCertSlotProperty is not set anymore, because it was never supported by Shill and leads to errors with Service::SetProperties.

BUG=234983, 126870
R=stevenjb@chromium.org

Review URL: https://codereview.chromium.org/22588002