InstantExtended: Nuke InstantOverlayController.
There is no longer an Instant overlay that needs controlling, so we can
remove InstantOverlayController. Mac's view code is warped around the
existence of an overlay, so cut out as much of the dead Instant code as
possible without removing OverlayableContentsController; I'm not a Cocoa
expert and don't know the best way to do that.

BUG=251262
TEST=None.
TBR=erg@chromium.org, rsesek@chromium.org, samarth@chromium.org

Review URL: https://codereview.chromium.org/22694002