Collapse one-time loop in PictureLayerImpl::ContentsResourceId().
This should not be a functional change, but it's clearer that this can
only loop one time by writing it without a 'for' statement. If this
should loop more than once, a functional change is required.

As described in http://www.viva64.com/en/b/0205/.

BUG=271533

Review URL: https://chromiumcodereview.appspot.com/22799006