aura: Remove CreateOffscreenContext from ui::ContextFactory.
Drop this method since it is not needed in the interface anymore. In
order to do this, make the cc::FakeContextProvider externally available
so ui::Compositor can make use of it for tests, and revamp it to make
it slightly more general.

Depends on: https://codereview.chromium.org/22293003/

R=jbauman, piman
BUG=258625

Review URL: https://chromiumcodereview.appspot.com/21026005