ash: Don't SetMasksToBounds on the workspace window.
If the workspace window masks to bounds, then it clips anything that
leaves the viewport. If you are tab-capturing a window, and move the
window partly outside the viewport, we should still be able to draw
and capture the window's contents. But this is prevented if we clip
the window to the viewport, as the compositor can not tell that the
window should not be clipped for the purposes of readbacks.

Since the ash workspace fills an entire compositor viewport, there
should be no need to clip anyways, as the compositor's output surface
is bounded at the same places.

R=ben@chromium.org, sky@chromium.org
BUG=263512

Review URL: https://chromiumcodereview.appspot.com/22638019