Fix bitmaps for our first synced notification service to override properly.
This is part of a change to allow a specific (Google chrome) bitmap to override
a generic (chromium) bitmap.

My first attempt tried to put it into a "common" directory, thinking that
src-internal would get copied over the files in "common" virtually at build
time.  However, more investigation revealed that instead the .grd file has a
switch and uses either a file from the "chromium" directory or the
"google_chrome" directory depending on which we are building.

This will place the ic_gplus_color_16.png bitmap in the right directory to be overridden by the chrome variant, and do the logic in the theme_resources.grd file to have the bitmap use the "chromium" version for chromium, and the "google_chrome" version for chrome.  The google_chrome version is being checked into the src-internals tree as part of a separate changelist (and will be checked in before this changelist).

BUG=263083
R=sail@chromium.org

Review URL: https://codereview.chromium.org/22982002