Fix DepthTextureTest.RenderTo gl_test.
It assumed that when reading a depth texture with an internal format such as DEPTH_COMPONENT, that the depth would be available in the red, green and blue channels. According to the ANGLE_depth_texture specification:

    "...The resulting <Tau> is assigned to <R_t>. In some implementations, <Tau> is
    also assigned to <G_t>, <B_t>, or <A_t>. Thus in table 3.12, textures with
    depth component data behave as if their base internal format is RGBA, with
    values in <G_t>, <B_t>, and <A_t> being implementation dependent."

So the green, blue and alpha channels are actually undefined. The test now only verifies that the depth can be read from the red channel.

According to the CHROMIUM_depth_texture specification:

"This extension is the same as ANGLE_depth_texture except it does not add support for TexStorageEXT."

According to WEBGL_depth_texture:

"As per the ANGLE_depth_texture specification, when a depth texture is sampled, the value is stored into the RED channel. The contents of the GREEN, BLUE and ALPHA channels are implementation dependent. It is therefore recommended to use only the r component of variables in GLSL shaders that are used to reference depth textures."

This is unavoidable because depth textures in D3D9, D3D11 and GL behave differently wrt the green, blue and alpha channels.

BUG=135229

Review URL: https://chromiumcodereview.appspot.com/22717003