[NaCl SDK] Simplify sdk_files.list so it breaks less often.
sdk_files.list previously included all files that examples and libraries used.
This was burdensome because we often didn't catch verification failures until
a trybot run.

This change sacrifices some correctness for simplicity; for all examples,
libraries and includes we only ensure that a directory exists (and that it has
at least one file in it).

This CL also fixes a bug I found in the verifier, where nested globs were not handled correctly.

BUG=none
R=nfullagar@chromium.org, sbc@chromium.org

Review URL: https://codereview.chromium.org/23100002