Remove dead code from webkit_support
webkit/support still has legacy code from when it was used to run layout
tests. This removes a large portion of code that did setup/teardown work
that was either only needed for layout tests and not webkit_unit_tests or
that was redundant with base::TestSuite.

R=darin@chromium.org

Review URL: https://codereview.chromium.org/22798008