Don't add a dash for empty description
In crrev.com/217109 I caused a regression that added a dash to the
omnibox popup even if the description was empty.

This CL adds back the original check for an empty description.

BUG=271951
TEST=Typed youtube.com in the omnibox. Verified that omnibox
results with an empty description didn't have a dash at the end.
NOTRY=true
R=shess@chromium.org

Review URL: https://codereview.chromium.org/23022003