cros/ash: Remove about:transparency tweak page
This is an old feature from back when we were deciding how transparent
the window captions should be.  The transparency values is not persisted
intentionally; users cannot use this as a setting feature.

Since we have long decided how transparent we want things to be,
rip this out.

BUG=270896
TEST=compiles, on Chrome OS about:transparency shows "This web page is not available"

Review URL: https://chromiumcodereview.appspot.com/22794011