Separate logout button from system tray
A logout button is shown in the ash launcher for public sessions. The
positioning of this button is problematic: It is part of the system tray,
causing users to accidentally click on the button (terminating the
session) when they meant to access the system tray bubble.

This CL moves the logout button out of the system tray. For screenshots
of the separated button, see: http://crbug.com/223846#c40

BUG=223845
TEST=Manual in VM

Review URL: https://chromiumcodereview.appspot.com/23099002