Fix a bug in got_XX_revision when a DEPS roll is involved with git-try/gcl try.
The function _handleGotRevision() is called *before* the patch is applied, so it
was inconditionally getting the stale revision. This CL fixes the problem for
try jobs triggered with gcl try or git try, it doesn't fix the issue with git cl
try or Rietveld triggered try jobs.

Add unit test. I wasn't sure if this part of the code was working for a new
repository I'm fetching so I unittested its output.

R=ilevy@chromium.org
BUG=

Review URL: https://chromiumcodereview.appspot.com/20800002