[sync] Fix focus issue in advanced sync settings dialog
When the user uses the keyboard to move between the "Sync everything"
and "Choose what to sync" options in the main drop-down menu of the
advanced sync settings dialog, the focus automatically gets transferred
over to the OK button instead of remaining with the drop-down menu.

This patch switches things around to assign focus to the correct element
only while originally displaying the dialog, and avoids the focus-stealing
that was going on when the drop-down menu choice changed.

BUG=272059
TEST=Use the keyboard to switch between items in the advanced sync dialog

Review URL: https://chromiumcodereview.appspot.com/23092003