aura: Allow in process ContextProvider to hold onscreen contexts.
This will allow us to use them in an OutputSurface. For the
common offscreen context case, we want to still share code
so add a CreateOffscreen() method to the context provider
so we don't have to replicate the WGC3D::Attributes decisions
around the codebase.

R=piman@chromium.org
BUG=258625

Review URL: https://codereview.chromium.org/23072008