CHR-458: Remove DCHECK which prevents from using custom ResourceBundle::Delegate
I need to create a Delegate for ResourceBundle. The only purpose of it is to provide our own fonts. Other resources should be handled by ResourceBundle. A method GetImageNamed() should also be handled by ResourceBundle. It works fine in release builds, but triggers an assert in debug builds.
DCHECK is triggered if a delegate does not create an image. This is wrong since the code can handle such case.

Review URL: https://chromiumcodereview.appspot.com/22418003