Expose GPU information to Telemetry.
The GPU on the system must be known to Telemetry in order for the test
expectations to be applied properly. Added SystemInfo domain to the
DevTools protocol and GetGPUInfo method to Telemetry's Browser class.
Added GPUInfo and GPUDevice classes to telemetry.core. Wrote and ran tests
for all new code.

BUG=245741
R=dtu@chromium.org,zmo@chromium.org,pfeldman@chromium.org,nduca@chromium.org

Review URL: https://chromiumcodereview.appspot.com/21682002