Prevent zoom bubble from closing on click
In crrev.com/217174 I caused a regression where clicking anywhere on the
zoom bubble would cause the bubble to close.

The problem was that clicking on the zoom bubble would remove focus
from the browser window. This caused
-[BaseBubbleController windowDidResignKey] to close the bubble.

Fix was to set the bubble's canBecomeKeyWindow flag to NO. 

BUG=273202
TEST=Zoomed a page. Clicked on the zoom deocration in the omnibox.
Pressed the plus button. Verified that the bubble didn't close.

Review URL: https://chromiumcodereview.appspot.com/23234002