Remove base::Callback from ContextProvider creation.
Currently a Callback is given to ContextProvider's creation methods
that is able to create a WebGraphicsContext3D and return it to the
provider. This callback is called synchronously during the create
method, and only one time.

Instead of requiring this indirection through a Callback, we can instead
just pass a WebGraphicsContext3D to the ContextProvider when it is
created, directly.

R=piman, sievers
BUG=258625

Review URL: https://chromiumcodereview.appspot.com/23247002