[Telemetry] Fix finding of minidump_stackwalk binary on mac.
The mac binary is in a subdirectory of the build directory, so it doesn't
work to simply take the dirname of the chrome executable. Instead, factor
out a FindSupportBinary() method that knows about the build directories.

We should use this for finding any other binaries like forwarder, md5sum, etc.

BUG=223572

Review URL: https://chromiumcodereview.appspot.com/23064005