telemetry: Unbreak --verbose command line flag
The --verbose flag was not working because logging.basicConfig() is a
no-op if any logging has been performed prior to calling it. As a side
effect of initializing the command line parser,
profiler_finder.GetAllAvailableProfilers() ends up initializing adb on
Android. Because this involves logging, setting the logging verbosity
after this had no effect.

This patch fixes the problem by not relying on logging.basicConfig() but
instead setting the logging level for the main logger directly.

Review URL: https://chromiumcodereview.appspot.com/22967004