Android WebView: build fixes for use_system_skia.
The skia shim headers were not being included at the front of the path
because of a gyp processing issue, causing it to pick up the bundled
headers instead by mistake; fix the gyp file.

Also, some of the chromium code depends on non-public headers; add the
required include path for this to work until we can fix the code.

BUG=263818,274425

Review URL: https://chromiumcodereview.appspot.com/23064007