Add support for modal dialogs in desktop aura.
This involves implementing the DesktopNativeWidgetAura::InitModalType function to
forward the call to the underlying DesktopRootWindowHost. 

On Windows this will make all modal dialogs window modal. We don't have support for
system modal dialogs in the current implementation.

This is on par with non aura windows.

BUG=249981
R=msw@chromium.org, robertshield@chromium.org, sadrul@chromium.org

Review URL: https://codereview.chromium.org/23171013