Native messaging support for service accounts.
This is an alternative implementation of https://codereview.chromium.org/19670021/ . This version performs the OAuth token exchange steps from inside the NativeMessaging host. This means that the client secret and the service account access token never leave the native messaging host.

Since the API keys are also needed to use the refresh token, this means that only the host binaries are able to get an access token for the service account.

BUG=224742

Review URL: https://chromiumcodereview.appspot.com/22769002