cc: refcount resources as we send them to the parent
This allows us to fully specify the resources used by the frame, so that we can:
- allow more than one frame in flight
- allow duplicating the frame for aura::Window::RecreateLayers

For every frame we generate, we increase the "exported_count" in the child.
Every time we import the resources for a frame, we increase the "imported count"
in the parent. The contract is that we return resources exactly once for every
time it's exported.

BUG=263068

Review URL: https://chromiumcodereview.appspot.com/23023005