Autoclosing of tray bubbles when the mouse is over the bubble but not moved by the user
The problem:

The user watches a full screen movie and uses the keyboard to change the volume. At that time the cursor is at the location where the bubble shows up.

The result is that the bubble will stay open until the user moves the mouse away.

What should happen instead is that as long as the user does not actively move the mouse, it should automatically close again.

The CL is looking at the Synthesize flag upon entry which indicates that this event was not a real user action. If that happens, the mouse will be monitored and if it is moved the event will be dispatched as a mouse entered event. Only when the entered event was sent, the exit event will be sent.

BUG=271121
TEST=visual

Review URL: https://chromiumcodereview.appspot.com/23200007