A status change of a launcher item did remove the ideal bounds of the item setting it to (0,0).
The problem arises by the fact that the view_model does contain the "ideal_bounds" on a per view basis and not on a "per slot" basis. When an item gets removed / added the "ideal_bounds" need re-calculation which is usually done by animating the launcher items to their correct location. In this case however the item was removed and re-inserted as a new item and the ideal location was lost.

I have to say that this scheme to manage the two data sets seems to ask fro trouble - even though I can see why it was done this way. Changing this "cleanly" would require a lot more refactoring which we possibly do not want to do at this time. 

BUG=273831
TEST=visual: Have a launcher item showing up automatically by launching. Pin and minimize to see.

Review URL: https://chromiumcodereview.appspot.com/23003012