Resubmit issue to fix issue where window bounds were being passed with wrong coordinates.
This is a speculative re-upload as the previous change was reverted with 
mysterious failures which seemed unrelated to my change.  It has been 
more than 1 week, and being unable to find any reason for the failures,
will try again with the same commit to see if the issues resurface.

Original review URL:
https://codereview.chromium.org/19681003/

BUG=176221
TBR=sky

Review URL: https://chromiumcodereview.appspot.com/22848003