Add a UUID based identification generator.
This CL adds a new identification generator that just uses a UUID
directly. It is stored to disk as a preference to ensure that it is the
same value across restarts.

The plan is to use this for session sync in the Chromium testshell.

BUG=272584

Review URL: https://chromiumcodereview.appspot.com/22980015