Android:Method to check if native library is initialized.
It's required by ContentProvider in Android to avoid loading native library
as a result of ContentProvider call. It allows ContentProvider to determine
if native library is initialized and proceed only if so. ContentProvider.query
method will drop the request if the library is not initialized.

BUG= 176000

Review URL: https://chromiumcodereview.appspot.com/23125004