[MediaGalleries] Move handling of MTP delegate from the refcounted object into the media file system registry.
This moves the coupling from the MG filesystem registry and the MTP service map into the registry file,
leaving the ScopedMTPDeviceMapEntry as a very thin class and hopefully clarifying the relationships.

I added a comment to the ScopedMTPDeviceMapEntry class that spells out current use. I think it could
potentially just be slurped into the registry file as an adjunct implementation class at this point.

R=thestig@chromium.org, vandebo@chromium.org

Review URL: https://chromiumcodereview.appspot.com/23260002