[Telemetry] Fix TestPerfProfile on some machines.
Previously the test depended on the result of perf record and would actually
run perf report. This didn't work because the perf record output had hardcoded
paths to the chrome binary on my machine.

Instead, depend on the output of perf report instead of actually running it.

BUG=273288

Review URL: https://chromiumcodereview.appspot.com/23245009