Make AccessibilityNodeData more compact.
The only goal of this tedious changelist is
less memory usage when storing the accessible tree.

The major changes are:
* Make almost everything a "sparse" attribute.
  For example, not every element has a name or
  value, so make them optional attributes rather
  than string fields that always take up memory.
* Use UTF-8 std::strings instead of string16s.
* Use vectors of pairs rather than maps.

If there is any slowdown after this change, we
can speed up lookup in vectors by sorting the
attributes and doing a binary search. There are
also opportunities to do more referencing and
less string copying.

When reviewing, look at the changes to the header
files first and make sure you understand them.
Everything else is just a consequence of those
changes.

BUG=96700

Review URL: https://chromiumcodereview.appspot.com/21269002