Fix crashes when NinePatchLayer doesn't have a resource.
First issue is that if NinePatchLayer doesn't get a ResourceId because
of memory pressure, it will DCHECK. This is a valid situation -- just
draw nothing (checkerboard would be okay as well).

Second issue is that NinePatchLayer passes the ResourceId in for the
texture in PushPropertiesTo, but does not call SetNeedsPushProperties
when painting a new resource. The documentation of Layer::Update
suggests that a side-effect of returning true from Layer::Update is
that SetNeedsCommit will be called ont this layer, but this is not
the case.

BUG=133097

Review URL: https://chromiumcodereview.appspot.com/22911024