cc: Don't require push properties for picture layers
PictureLayer now only needs push properties if its pile has changed. In
order to make this possible, some of the non-pushing post-commit work
had to be deferred so that every layer that got committed (even if it
didn't get pushed) would go through it. This is lazily done with a
boolean flag in PictureLayerImpl, with some DCHECKs for sanity about its
state.

As every layer on the pending tree that will get drawn will be
CalcDrawEtc'd and get UpdateTilePriorities first, this seems a safe
place to put this lazy check.

This is a re-land of r214533 combined with r216218 and depends
on the bugfix https://codereview.chromium.org/23295003/.

R=danakj@chromium.org
BUG=259402

Review URL: https://chromiumcodereview.appspot.com/23129009