Add --stable-release-mode to content_shell
The new flag is used to force testing content_shell with only stable
flags enabled, which will enable testing what is exposed in stable.

As part of this change, moved the "test" runtime features enabling
into ShellRenderProcessObserver. This is a good step as TestInterfaces
(that handles this now) doesn't know about which flags are passed on
the command line.

The change is not testable right now as it requires some changes in
TestInterfaces in Blink to avoid enabling all the flags.

BUG=234853

Review URL: https://chromiumcodereview.appspot.com/22799014