Add logging to diagnose cursor visibility issues on Pixel
Add logging to determine if
CompoundEventFilter::SetCursorVisibilityOnEvent() ever keeps
a currently-invisible cursor invisible when it is called
with |show| = true. I suspect this as a possible cause of
issues 248231 and 267789.

LOG(ERROR) was used instead of LOG(WARNING) because I am
particularly interested in what happens at the login
screen, and the logs in /var/log/ui/ do not log WARNINGs.

I added IsCursorLocked() to the interface for
CursorManager in order to prevent my new logging from
appearing while in pointer lock (which is a case I do not
care about).

BUG=275826

Review URL: https://chromiumcodereview.appspot.com/22867023