Hide all elements marked as "gsm-only" in internet-detail.html.
I discovered that certain elements that should be hidden for CDMA
networks have the CSS class "gsm-only". These elements (such as the
APN selection UI and the "Security" tab) should be hidden for non-GSM
networks. Updated the code so that such elements are hidden. Also updated
all settings entries that need to be conditionally hidden based on
"Cellular.Family" to belong to the "gsm-only" and "cdma-only" classes.

BUG=264097
R=stevenjb@chromium.org

Review URL: https://codereview.chromium.org/22871013