Refactored privet http API to add context to delegate
Added pointers to operations to all privet HTTP delegates, a pointer to the
parent HTTP object to all privet operations, and a name to all privet HTTP
clients, to allow users a better grasp on the privet HTTP context.

This is mostly a precursor to notifications, where we'll need potentially
multiple info operations at the same time.

BUG=245385

Review URL: https://chromiumcodereview.appspot.com/22407014