Always treat the app list shim process as a "toggle" for the OSX App Launcher.
Since the OSX App Launcher is drawn as a UI Element, there is no need
for the shim process to stay running to indicate that the UI is active
with a running light in the dock.

Switching to a toggle simplifies the logic to handle both shim-activated
launchers and webstore-activated launchers. Closing the shim process
immediately simplifies object lifetime, prevents OSX intercepting a
"relauch" and generating a different codepath, and better reflects the
Dock state imposed by using the UI Element launch flag (i.e. no running
light).

With this CL, the app list shim process is closed with
APP_SHIM_LAUNCH_DUPLICATE_HOST which indicates a normal (i.e. non-error)
situation telling the shim that it should close, because there is
already a mechanism running that handles the request.

BUG=276157,276085,261963
TEST=Get the OSX App Launcher by installing an app from the webstore.
Wait for it to complete and install another app - app lanucher should be
shown. When complete, clicking the dock icon should dismiss the app
launcher.
TEST=Have OSX app launcher enabled. Open Chrome browser. Focus a
non-Chrome OSX app on top of the browser. Open App launcher via dock
icon and dismiss it via dock icon. Focus should stay with the non-Chrome
app, and browser window should not be raised.

Review URL: https://chromiumcodereview.appspot.com/22865022