Cherry-pick scheduling fix to uniforms/uniform-samplers-test.html
This patch cherry-picks an upstream fix to
uniforms/uniform-samplers-test.html to make it execute in smaller slices
instead of running the entire test within a single function. This is
needed because on devices like the Nexus 7, the monolithic test blocks
the main JavaScript thread for over a minute and this interferes with
test automation.

BUG=272080

Review URL: https://chromiumcodereview.appspot.com/23060009