Remove unnecessary [AppsGridViewItem onInitialModelBuilt] from OSX App List.
This is a partial revert of r216090. onInitialModelBuilt would allow
parts of an initial model to be represented in the UI which were set
while the model was not being observed. However, the model itself is
only updated when _its_ observers are registered, which happens at the
same time as the things that would observe the model are created.

The correct fix was r216954 which registers the things updating the
model earlier so they can observe changes and update the model.

This CL removes the onInitialModelBuilt function from AppsGridViewItem
and related test - HighlightedOnFirstShow - whose interesting bits are
still covered by tests AppsGridControllerTest.ItemInstallProgress and
AppsGridControllerTest.EnsureHighlightedVisible

BUG=267814
TEST=No regressions. Specifically, after installing the _first_ app from
the webstore and receiving the app launcher, the app launcher should not
pop up. Then, clicking the dock icon should show the just-installed app
highlighted.

Review URL: https://chromiumcodereview.appspot.com/22813012