Don't pass an unnecessary bool along the chain of "update toolbar" calls.
All we really need is the other arg, the WebContents* of the tab to update,
because in the end, that's all the location bar classes take; before then the
bool is handled by forcing the pointer to NULL if the bool is false.  So just do
that as early as possible.

BUG=none
TEST=none
R=sky@chromium.org

Review URL: https://codereview.chromium.org/22903018