[NaCl SDK] nacl_io: html5fs should return EEXIST on mkdir("/")
It seems the Pepper returns PP_ERROR_NOACCESS (which is translated to EACCES)
when you try to create the root directory, instead of PP_ERROR_FILEEXISTS as I
had expected.

This causes libtar extract (for example) to fail when extracting to the root of
the html5fs mount.

BUG=none
R=sbc@chromium.org

Review URL: https://codereview.chromium.org/23129022