Currently, Chromium does not listen for error notifications from QTKit, i.e. they are ignored.  QTKit device creation can succeed, and then later fail asynchronously via such a message, never producing any frames.  If Chromium doesn't listen for the error notifications, then this results in a getUserMedia call that never fires its success or failure callbacks.
For example, with this change, if too many cameras are connected to the same USB hub, then bandwidth limitations may cause some of the cameras to fail to open.  This results in a QTCaptureSessionRuntimeErrorNotification whose localized text reads "The operation couldn't be completed. (OSStatus error -536870164.)".  With this change, that notification results in a getUserMedia failure callback to the javascript (with event.name == 'PERMISSION_DENIED', which is the only currently supported error code).

BUG=265704

Review URL: https://chromiumcodereview.appspot.com/22982008