Fix NULL dereference in StatusTrayWin.
It seems possible for window messages to be queued to status tray
icons while Chrome is processing their removal from the StatusTray,
causing WndProc to be called with an invalid status icon id in
the wparam.  Now we'll return early if the status icon id is not
found.

BUG=275823

Review URL: https://chromiumcodereview.appspot.com/23079011