[sync] Properly handle case where sync is disabled by dasher admin
If a user signs in with a dasher account for which sync is disabled by
admin, SyncerProtoUtil::PostClientToServerMessage neglects to check if
sync is disabled, and incorrectly flags the server response as a
NOT_MY_BIRTHDAY error. This results in incorrect sync state showing up
in about:sync, and UI brokenness, where we mistakenly expose the "Set up
sync..."  button.

This patch updates SyncerProtoUtil::PostClientToServerMessage to first
check the server response to see if sync is disabled, prior to checking
for other sync protocol errors.

BUG=276030
TEST=Sign in with a dasher account for which sync is disabled. The "Set
up sync..." button must not be exposed in the UI.

Review URL: https://chromiumcodereview.appspot.com/23030008