Reload font setting at the bootstrap of the OOBE.
If the ApplicationLocale in preference is empty, OOBE will be start with
initial_locale in VPD value. In that case we need to reload font settings with
initial_locale, otherwise wrong font will be used. This is problematic in
Japanese, Chinese font will be used without this reload.

BUG=127223
TEST=Manually confirmed with power-washing.

Review URL: https://chromiumcodereview.appspot.com/22858005