Add translation of Shill's security type "wpa" to ONC.
In certain cases, Shill reports the security type ( flimflam::kSecurityProperty) of a WiFi endpoint. That type can be flimflam::kSecurityWpa and isn't handled correctly by the translation to ONC, which in turn is used by the networkingPrivateAPI and the network policy implementation.

This commit adds the missing translation.

BUG=277417

Review URL: https://chromiumcodereview.appspot.com/23390002