Land Recent QUIC changes.
Removing a check causing the encrypter to return NULL when packets were
serialized out of sequence number order and added NULL checks where the
encrypter is used to ensure if this occurred again, it would not be a
fatal error.

Merge internal change: 51212615

Removing a DCHECK and associated comments about looping forever.

Merge internal change: 51211309

A regression test for the gmail multi-stream timeout bug.

Merge internal change: 51201046

Fixed the tests in CL 51179724

Merge internal change: 51185931

Q009: now sending a uint32 priority as the first data on a stream.
This priority is not yet used: it's stashed in the stream and ignored.

Merge internal change: 51179724

R=rch@chromium.org

Review URL: https://chromiumcodereview.appspot.com/23005020