Merge internal CL: 51214421
To make reasoning about crypto frames easier, change
QuicConnection::SendStreamData to not combine crypto frames with any
other frames in a single packet.

QuicServer::Shutdown should close the fd so that the QuicServer can be
restarted and bind to the same port.
        
This CL was written in collaboration with Jim Roskind.

R=jar@chromium.org,rch@chromium.org
BUG=none
TEST=net_unittests

Review URL: https://chromiumcodereview.appspot.com/23301016