Change BluetoothDeviceList to single selection model and set the device to be selected when "x" is clicked on the item.
The multiple selection model allows user to select multiple devices to pair/connect/disconnect/unpair, which is confusing and inconsistent with the real work flow, change it to single single selection model.

Always select the device that has been clicked on "x", which is much easier to understand and consistent with "Connect" button state.

BUG=270197
TBR=jhawkins

Review URL: https://chromiumcodereview.appspot.com/23220008