[sync] Force re-auth during sync setup if Oauth2 token is unavailable
If an enterprise user disables sync via a dashboard clear, and goes on
to re-enable sync at a later time, set up can fail due to an expired or
missing Oauth2 token.

This patch updates SyncSetupHandler::OpenSyncSetup to first check if an
auth token is available for sync, and if not, force a re-auth.

R=atwilson
TBR=fgorski,rogerta
BUG=276650
TEST=Disable sync for an enterprise user, exit chrome, delete "Web
Data*" from the profile dir, restart chrome, and set up sync. Re-auth
must be forced in this case.

Review URL: https://chromiumcodereview.appspot.com/23005016