This make the GN current dir be set on thte scope. This fixes a bug where a template execution uses the directory of the file containing the template declaration. This makes it impossible to refer to the input files which are normally specified relative to the invoking file.
I added a defined() function that returns true if a variable has been defined. This needed some magic function invocation help because we don't actually want to evaluate the argument.

I added an optional second argument to assert for an error message.

We now tolerate absolute file paths for sources and outputs. Grit was outputtings these, which was confusiong some of our assertions. We'll now check if such locations are inside our source tree and fix them up accordingly.

I fixed a bug in the "desc" command where a target would be listed as a dependent of itself. I also added the target type to the output.

Note that I didn't implement MakeAbsolutePathRelativeIfPossible on Windows. I will do this on Windows in a second pass. Currently it's an #error.

There are several fixes to custom script target outputs and fixes to variables leftover from when I converted paths to start with two slashes.

BUG=
R=scottmg@chromium.org

Review URL: https://codereview.chromium.org/23396002