Native messaging installer for Mac.
Since Packages doesn't allow for a different source and destination file name, I decided to make the build create the native messaging manifest under the name it eventually has under Linux, which entailed a small change to the Linux installer as well.

One thing I'm not entirely happy about with this CL is that the Mac is installed as org.chromium.chromoting, whereas the native messaging component is called com.google.chome.remote_desktop. It would be nice if we could use the same name for both. For this CL, I think keeping the two different names is best for the sake of keeping it small.

BUG=173509

Review URL: https://chromiumcodereview.appspot.com/22901019