Fix a problem that certain videos do not display
This is a regression introduced by https://codereview.chromium.org/23137013/
Previously if the requested field is NULL, we will throw an exception and return. The success value is false.
However, the new behavior will return a true success code.
This causes the video size to be incorrectly set in WebMediaPlayerAndroid.

TBR=yfriedman
BUG=277105

Review URL: https://chromiumcodereview.appspot.com/22839019