cc: Remove reference counting from PrioritizedTileSet.
This removes reference counting from tile management and instead
lazily constructs the PrioritizedTileSet as needed without
ManageTiles() having to be called. A ManageTiles() call will force
the reconstruction of the TileSet.

This doesn't change the current behavior as ManageTiles() is
called whenever the TileSet is changed in a meaningful way. We can
enforce this using DCHECKs but that makes consolidating
ManageTiles() calls harder.

This gives a ~125% performance improvement to
TileManagerPerfTest.ManageTiles.

BUG=268441
TEST=cc_unittests,cc_perftests

Review URL: https://chromiumcodereview.appspot.com/22825029