Fixes possible use after free in WrenchMenu.
Specifically ZoomView installs a listener on the HostZoomMap in the
constructor and removes it in the destructor. To find the HostZoomMap
ZoomView access menu_. Problem is in some situations menu_ may have
been destroyed already. Fix is to add an Observer and do cleanup when
WrenchMenu is destroyed (or in the destructor).

I would have liked a test for this, but that currently requires an
interactive test which I'm loathe to write.

BUG=236581
TEST=with aura build bring up wrench menu, and click close button on
     browser. Make sure don't crash.
R=skuhne@chromium.org

Review URL: https://chromiumcodereview.appspot.com/23187006