Create tryserver appengine app.
This appengine app is a step towards decentralized job control. It accepts POSTs
of json blobs describing buildbot jobs to its 'push' endpoint. Then, requests to
the 'pull' endpoint will be served not-yet-picked-up jobs. The poller can then
POST to the 'accept' endpoint to indicate that it has successfully picked up and
started a job. The 'peek' endpoint serves to inspect the job queue without
marking any jobs as potentially taken.

This is intended to interact with 'git try' on the client side to push blobs to
the app, and a new buildbot poller and scheduler to consume blobs and start
jobs.

R=cmp@google.com

Review URL: https://codereview.chromium.org/23093011