stop setting the initial duration to 100 seconds
Currently we set the initial default duration to 100 seconds so that user can perform seek if media is not prepared. This happens when user is not on a wifi network in order to save data consumption.
However, this comes with an ugly hack that we need to recalculate the seek position when duration becomes known.
Removing the initial duration hack will still allow user to play the media in non-wifi environment.
However, user cannot seek initially until the duration is known.
When wifi is available, MediaResourceGetter can get the media duration by itself. Otherwise, hitting the play button will cause the player to prepare itself, and extract the duration from the media url.

BUG=276693

Review URL: https://chromiumcodereview.appspot.com/23112024