Use NativeWindow instead of anchor view to get icon scale in app list view.
The app list with the alternate shelf layout does not have an anchor, so the
scale factor for preloading icons is always set to the default. On link, this
causes the icons to be loaded twice, resulting in a noticeable delay or flicker
whenever the app list is opened.

Instead of the anchor, we will use the native host window directly.

BUG=274046
TEST=manual, set alternate shelf layout in about:flags on link and reboot, press search key, icons paint immediately with app titles

Review URL: https://chromiumcodereview.appspot.com/23038022