Add a function to change the mouse cursor icon set to aura/ash.
This patch implements a function to change the asset set of mouse cursor icons, and adds the method of changing the cursor set into CursorManager.

This feature will be used with Large Cursor feature. Currently, it just enlarges the normal cursor assets. This patch makes it use the large-size cursor assets instead. The large assets will be added in an another patch after this patch is committed.

BUG=247254
TEST=manuallly. Tests pass.

Review URL: https://chromiumcodereview.appspot.com/22319004