History: Clear Browsing Data should not overscroll along with history list.
On iOS and Android, the Clear Browsing Data button is fixed at the bottom
of the screen. When scrolling the list of history entries, the scrollbar
overlaps the button, and when you hit the bottom of the list, the button
overscrolls along with the history entries.

The solution is to add a separate scrolling container that does not overlap
with the Clear Browsing Data button. Unfortunately, this means we lose the
overscroll effect on Android (it no longer supports the -webkit-scrolling
property).

BUG=262219

Review URL: https://chromiumcodereview.appspot.com/23241025