Pepper: Fix FileSystemBrowserHost for isolated fs.
Currently the PepperFileSystemBrowserHost doesn't have a sane filesystem
context for isolated file systems. This was causing some crashes in an
application that used a CRX filesystem, which caused the FileRef change to be
reverted.

This was hard to debug because it was causing a browser crash; I cleaned up the
code in PepperInternalFileRefBackend to behave better when there's no file
system context.

TBR=jschuh
BUG=225441

Review URL: https://chromiumcodereview.appspot.com/22867028