Cleaned up the Overlapped Reads field trial where we compared the existing method for reading from sockets on Windows (using overlapped I/O) with using non-blocking I/O. This was the "OverlappedReadImpact" field trial that started in November 2012 and ended in June of 2013.
The original patch for implementing the trial was here:
https://chromiumcodereview.appspot.com/10916016

The non-blocking I/O code path resolved the issue we were
seeing in lab testing but did not show a marked improvement
in the field (nor did it show a degradation).  The results
for 3 months of data from the field trial showed effectively
no change in the page load time or HTTP Job times
(differences are within the noise variation).

Since the non-blocking code path is a little cleaner,
doesn't rely on growing buffer sizes and fixed the issue we
were seeing in the lab, the decision was made to keep the
non-blocking implementation and eliminate the overlapped
read code path.

Review URL: https://chromiumcodereview.appspot.com/22188002