Fix overwriting other data in InitBrowserCrashDumpsRegKey.
This was causing random crashes depending on the layout of the linked objects.
On Windows in Debug, swprintf appears to always write 'fe' to all bytes which
could be written to. The second parameter of swprintf is the number of wide
characters, not the number of bytes.

R=cpu@chromium.org, rogerm@chromium.org

Review URL: https://codereview.chromium.org/23300006