[NaCl SDK] Fix examples/api/core test flake on Host builds.
Host builds (i.e. trusted plugins) of the examples were not waiting for modules
to load before running tests. This is because a trusted plugins do not send
progress events.

The host builds were still dispatching a fake 'load' event (after an arbitrary
time of 100ms), so moduleDidLoad was being called. This could happen after
the tests started to run. If so, the common.naclModule variable would not be
set when the tests were being run, which would cause the
common.naclModule.postMessage() call to fail.

The fix is to send more fake events: 'loadstart', 'load', 'loadend', and set
the readyState variable correctly. Now, before running the tests, we can wait
for host plugins in the same way we wait for NaCl modules.

BUG=none
R=sbc@chromium.org

Review URL: https://codereview.chromium.org/23471002